+++ lib/carrierwave/sanitized_file.rb
@@ -1,12 +1,6 @@
 require 'pathname'
 require 'active_support/core_ext/string/multibyte'
-
-begin
-  # Use mime/types/columnar if available, for reduced memory usage
-  require 'mime/types/columnar'
-rescue LoadError
-  require 'mime/types'
-end
+require 'mini_mime'
 
 module CarrierWave
 
@@ -267,9 +261,14 @@ module CarrierWave
     def content_type
       return @content_type if @content_type
       if @file.respond_to?(:content_type) and @file.content_type
-        @content_type = @file.content_type.to_s.chomp
+        if @file.content_type.respond_to?(:content_type)
+          @content_type = @file.content_type.content_type.chomp
+        else
+          @content_type = @file.content_type.to_s.chomp
+        end
       elsif path
-        @content_type = ::MIME::Types.type_for(path).first.to_s
+        mime_type = ::MiniMime.lookup_by_filename(path)
+        @content_type = (mime_type && mime_type.content_type).to_s
       end
     end