+++ lib/carrierwave/uploader/download.rb
@@ -18,9 +18,9 @@ module CarrierWave
 
         def original_filename
           filename = filename_from_header || filename_from_uri
-          mime_type = MIME::Types[file.content_type].first
+          mime_type = MiniMime.lookup_by_content_type(file.content_type)
           unless File.extname(filename).present? || mime_type.blank?
-            filename = "#{filename}.#{mime_type.extensions.first}"
+            filename = "#{filename}.#{mime_type.extension}"
           end
           filename
         end