+++ spec/sanitized_file_spec.rb
@@ -1,11 +1,5 @@
 require 'spec_helper'
-
-begin
-  # Use mime/types/columnar if available, for reduced memory usage
-  require 'mime/types/columnar'
-rescue LoadError
-  require 'mime/types'
-end
+require 'mini_mime'
 
 describe CarrierWave::SanitizedFile do
   before do
@@ -189,7 +183,7 @@ describe CarrierWave::SanitizedFile do
 
     it "handles Mime::Type object" do
       file = File.open(file_path('sponsored.doc'))
-      allow(file).to receive(:content_type).and_return(MIME::Type.new("application/msword"))
+      allow(file).to receive(:content_type).and_return(MiniMime.lookup_by_content_type("application/msword"))
       sanitized_file = CarrierWave::SanitizedFile.new(file)
       allow(sanitized_file).to receive(:file).and_return(file)